@@ -3,6 +3,7 @@
   "version": "1.0.2",
   "description": "League of Legends Patch Notifier - Automatically monitors LoL patch notes and sends Discord notifications",
   "main": "dist/app.js",
+  "type": "module",
   "scripts": {
     "start": "node dist/app.js",
     "dev": "tsx watch src/app.ts",
@@ -31,7 +32,11 @@
     "release:win:major": "pwsh -ExecutionPolicy Bypass -File ./Scripts/release.ps1 major",
     "release:win:dry-run": "pwsh -ExecutionPolicy Bypass -File ./Scripts/release.ps1 patch -DryRun",
     "changelog": "git log --oneline --decorate --graph --since=\"1 month ago\"",
-    "version:check": "node -p \"require('./package.json').version\""
+    "version:check": "node -p \"require('./package.json').version\"",
+    "test:run": "jest --passWithNoTests",
+    "quality:check": "npm run lint && npm run format:check && npm run build && npm run test:run",
+    "quality:fix": "npm run lint:fix && npm run format && npm run build && npm run test:run",
+    "prepare": "husky"
   },
   "keywords": [
     "league-of-legends",
@@ -46,33 +51,46 @@
   "author": "",
   "license": "MIT",
   "engines": {
-    "node": ">=18.0.0",
+    "node": ">=20.0.0",
     "npm": ">=8.0.0"
   },
   "dependencies": {
-    "@google/generative-ai": "^0.24.1",
+    "@google/generative-ai": "^0.21.0",
     "axios": "^1.11.0",
     "cheerio": "^1.1.2",
     "dotenv": "^16.4.5",
-    "node-cron": "^4.2.1",
-    "winston": "^3.17.0"
+    "node-cron": "^3.0.3",
+    "winston": "^3.15.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.15.0",
     "@types/jest": "^29.5.0",
-    "@types/node": "^20.16.17",
+    "@types/node": "^22.10.0",
     "@types/node-cron": "^3.0.11",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
-    "eslint": "^8.57.0",
+    "@typescript-eslint/eslint-plugin": "^8.17.0",
+    "@typescript-eslint/parser": "^8.17.0",
+    "@vitest/coverage-v8": "^2.1.8",
+    "eslint": "^9.15.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.13.0",
+    "husky": "^9.1.7",
     "jest": "^29.7.0",
+    "lint-staged": "^15.2.11",
     "nodemon": "^3.1.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.4.1",
     "rimraf": "^5.0.10",
     "ts-jest": "^29.2.0",
     "ts-node": "^10.9.2",
-    "tsx": "^4.19.1",
-    "typescript": "^5.5.4"
+    "tsx": "^4.19.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.17.0",
+    "vitest": "^2.1.8"
+  },
+  "lint-staged": {
+    "*.ts": [
+      "eslint --fix",
+      "prettier --write"
+    ]
   }
 }
