@@ -7,7 +7,7 @@ import path from 'path';
 import { FileStorage } from '../utils/fileStorage';
 import { Logger } from '../utils/logger';
 import { config } from '../config';
-import { PatchNote, AppState, AppError } from '../types';
+import { AppError, type AppState, type PatchNote } from '../types';
 
 export class StateManager {
   private readonly stateFilePath: string;
